2/8/2024
This exercise using the Bing Liu lexicon.
1.importing Libraries :
To manipulate data, import Pandas as pd.
For jobs involving natural language processing, NLTK is imported.
It imports the preprocessing module from the sklearn package.

2.Loading the dataset
A DataFrame called df is filled with the dataset from the CSV file "Arts_Crafts_and_sewing_5.csv".

3.Obtaining the Bing Liu vocabulary:
One downloads the NLTK opinion_lexicon. Both positive and negative words used in sentiment analysis can be found in this vocabulary.

4.Establishing the Dictionary:
From the Bing Liu lexicon, two dictionaries are made: one for positive words and the other for negative terms. A sentiment score of 1 is given to each word for good words and -1 for negative terms.

5.Determining the Meaning of Sentiment Analysis:
The sentiment score is determined by adding the sentiment scores of each individual word in the Bing Liu lexicon. The function bing_liu_score(text) takes an input text, tokenizes it into words, then transforms them to lowercase.

6.Using the Function for Sentiment Analysis:
The DataFrame's "text" column is subjected to the sentiment analysis function, and the sentiment scores are recorded in a new column called "Bing_Liu_Score."

7.Code of Output:
'overall', 'text', and 'Bing_Liu_Score' are the columns that are printed on the dataframe.

Furthermore, the groupby function is utilized to determine the mean sentiment score for every distinct value in the 'overall' column.