assignment1/29/2024
It appears that this script is an analysis pipeline for taking data out of the Reddit "Hockey" subreddit and evaluating it. Let's examine the functions of each script section in detail:

Data Collection: The script begins by importing the required libraries and use the PRAW library to authenticate itself with Reddit's API. It retrieves content from the "Hockey" subreddit on trending themes.

Data Cleaning: The program strips stopwords, Unicode characters, and emojis from the titles of the topics it has fetched. For this, it makes use of NLTK's stopwords corpus and regular expressions.

Preprocessing and Tokenization: The cleaned titles are tokenized, and stopwords are once more eliminated. By taking this step, you can be sure that the text data is in an analysis-ready format.

Sentiment Analysis: Using TextBlob, the script analyzes the titles for sentiment and determines their subjectivity.Understanding the sentiment conveyed in the titles is made possible by this approach.


Data Visualization: Lastly, the script uses the matplotlib and seaborn libraries to visualize the data. It creates histograms to illustrate the sentiment score distribution and bar charts to demonstrate the distribution of titles. It also generates a word cloud to show the most frequently occurring words in the titles graphically.


Opinion Lexicon Analysis: It analyzes the sentiment of the titles using NLTK's Opinion Lexicon. For every title, this approach generates a sentiment score that is both positive and negative.

Overall, this script serves as a comprehensive pipeline for collecting, cleaning, analyzing, and visualizing data from the "Hockey" subreddit, providing valuable insights into the topics and sentiments expressed within the subreddit.